Improved algorithm
I've improved your algorithm, made the code much shorter and fixed the bugs. This is an easy implementation of the insertion sort: it starts iterating from the second number of the list, then finds all the previous number that are bigger than the current element and move them one position ahead, and so on for all the elements of the list, sorting the numbers in the right order in the end.